Change how link works
Link previously assumed we would only want to be linking internal
projects, an assumption that might not be suitable for oss use.

Rather than trying to use internal namespaces, we now use a qualified
project name and link it to a folder.

We assume that we link to ../project-dir, it may be useful to
parameterize this